+++ Logos/Changes.md
@@ -1,5 +1,10 @@
 # Utopia Logos - List of Changes
 
+## v0.5 (in development)
+### Breaking changes
+- **UniqueTextPlacementAccessLike** now extends **UniqueModelAccess**, which changes the abstract property from 
+  `accessCondition: Option[Condition]` to `condition: Condition`
+
 ## v0.4 - 23.01.2025
 This update changes how words are stored in the database, making this a rather breaking update when compared to v0.3. 
 Besides changing this storage logic, this update adds some previously missing interfaces / features.